`size_t` in `mojo::DataPipe[Consumer|Producer]Handle`: /extensions.
This CL makes use of the new `size_t`-based APIs in the remaining
directories like `//components`, `//extensions`, or `//storage`.
See https://crrev.com/c/5414148 for motivation.

This CL may have introduced new integer conversions - in those cases the
CL attempts to have well-defined behavior (e.g. using
`base::checked_cast`).  OTOH, this CL did *not* attempt to fix old
integer operations (e.g. old `int` => `uint32_t` / `size_t` conversions,
or old additions that may potentially overflow).

This CL was uploaded by git cl split.

R=mattreynolds@chromium.org, reillyg@chromium.org

Bug: 40284755
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5462185
Commit-Queue: Matt Reynolds <mattreynolds@chromium.org>
Auto-Submit: Łukasz Anforowicz <lukasza@chromium.org>
Commit-Queue: Łukasz Anforowicz <lukasza@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1288842}

+++ extensions/browser/api/serial/serial_apitest.cc
@@ -179,7 +179,7 @@ class FakeSerialPort : public device::mojom::SerialPort {
 
   void DoWrite(MojoResult result, const mojo::HandleSignalsState& state) {
     const void* data;
-    uint32_t num_bytes;
+    size_t num_bytes;
 
     if (result == MOJO_RESULT_OK) {
       result = in_stream_->BeginReadData(&data, &num_bytes,
@@ -243,7 +243,7 @@ class FakeSerialPort : public device::mojom::SerialPort {
     out_stream_watcher_.ArmOrNotify();
   }
 
-  void WriteOutReadData(uint32_t num_bytes) {
+  void WriteOutReadData(size_t num_bytes) {
     MojoResult result = out_stream_->WriteData(buffer_.data(), &num_bytes,
                                                MOJO_WRITE_DATA_FLAG_NONE);
     if (result == MOJO_RESULT_OK) {
